Voters' votes can be influenced by past elections
After each election, a voter has an 80% chance of storing the results
in their "memory". 80% is used here because, in the real world, people
may not follow certain elections or they'll think certain elections are
rigged.

Then in the next election, each voter's party preference order
has a 25% chance of being rearranged based on the results in their memory.
This makes voters of less popular parties vote for a larger party that sill
aligns with their views. I think this needs to be tweaked to take into account
how much the alignment with the most popular they are okay with differs from
their favorite party.

Still a little bit of work to do but this is showing promising results.

Signed-off-by: Jadon Fowler <j@jadon.io>